+++ app/date_parser.py
@@ -2,7 +2,9 @@
 from datetime import datetime, timezone
 
 
+# see: https://docs.python.org/3/library/datetime.html#strftime-strptime-behavior
 DATE_FORMAT = "%Y-%m-%d %H:%M:%S.%f%z"
+
 
 class DateParser:
     """Mixin for date parsing / interfacing with google sheets date formatting."""
@@ -15,13 +17,14 @@ class DateParser:
         return datetime.now(tz=timezone.utc)
 
     @staticmethod
-    def parse_timestamp(ts:str):
-        """Converts a timestamp string to a datetime object.
+    def parse_timestamp(ts:str) -> datetime:
+        """Converts a timestamp string to a datetime object as necessary.
+            Ensures you are working with a datetime object.
 
             Params:
                 ts (str) : a timestamp string in format provided by google sheets
 
-            Example: SpreadsheetService.parse_timestamp('2023-03-08 19:59:16.471152+00:00')
+            Example: DateParser.parse_timestamp('2023-03-08 19:59:16.471152+00:00')
 
             Returns a datetime object.
         """
